ICMSLST-1467 Use Django's mail backend for SMTP connections
Instead of having the MailServer class handle opening and closing SMTP
connection, this change uses the Django mail backend stuff for sending
messages. This requires passing an instance of EmailMessage to the
send function, which wraps a MIMEMultipart message.

The tests were updated because Django always adds headers for the
sending date and a random message ID. For now the tests still use an
SMTP server (MailHog) to send a message and check what sent, but this
will be removed in the future.

N.B. I have preserved the behaviour where we end up with both
`Content-Transfer-Encoding: base64` and `Content-Transfer-Encoding: 7bit`
as headers, with the payload as plain text. But we should fix this.