Remove manual scheduling of task to send licence details to HMRC
We have a periodic task send_licence_details_to_hmrc() that runs
every 10 min to see if there are any LITE issued licence details that
need sending to HMRC. But under some conditions this is triggered
manually.

After we relay a reply back to SPIRE we are scheduling this manually.
The intention is to avoid a potential situationn where sending of
LITE licence details can get delayed without this.

As we know we can only send one message at a time until it is
fully processed (i.e., status updated as reply_sent).

So a possible scenario where,

 1. We are waiting for a reply for SPIRE message (status = reply_pending)
 2. Licence issued in LITE but we cannot send this just yet
 3. Reply comes and we relay back to SPIRE.
    At this point we can send LITE details but the task may not be ready to run yet.
 4. We get another message from SPIRE and say manage_inbox() task managed
    to run so this will end up sending another SPIRE message and we
    again wait for reply (back to first step)

Because of this change we would've picked LITE details at
the end of 3 and sent to hmrc.

But we noticed that the condition is not correct.
If we receiver is SPIRE then the extract_type would be 'licence_reply' as
we only send licence reply and usage data mails back to SPIRE and never
a licence data (these go to HMRC).

So it is safe to assume that this has never been executed and the above
potential scenario might not be an issue. Even if that happens it will
only delay sending of LITE details but when SPIRE traffic reduces they
get picked up.

If this is really an issue and LITE details are not being picked up then
healthcheck starts failing and we will come to know about this.

Considering all of the above it is decided to remove this condition.
This commit serves as the reasoning behind this decision.