Do not retry sending lite licence details if there is validation error
The task that send LITE licence details is a periodic task that runs
every 10 min. Currently it is configured to retry on validation error
also which is redundant as it is a periodic task.

Besides if there is a validation error then it is most likely to fail
till that is fixed (usually code change) and there can be another
retry instance when the task runs in the next cycle again.