Remove healthcheck item that checks licence updates task responsiveness
Previously we were using background_tasks which is not a standard
task scheduler but now we use Celery so we can be a lot more confident
on beat scheduler.

We also have another check if the payloads are being processed or not.
If this task is not scheduled as expected (by celery) then the other
check would fail so we are not compromising on health checks.

Also the advantage would be that the health check is outcome based
and not rely on the implementation details.