Add helper function to acquire lock before a critical section
Locking is achieved by adding a key to the cache with an expiry time.
If the key doesn't exist then add it and acquire lock.
Subsequent attemtps to add the key fails (no lock) hence they have
to wait (till it is released or until expiry).

Default expiry is 10 min.

Co-authored-by: Le H Ngo <hgle.ngo@gmail.com>