Add task to send emails
Add new task that is dedicated to sending emails.
We have a need to send email from different tasks. When processing
a reply if the any of the licences are rejected then we schedule
a task to notify users of these rejected licences. In this situation
we noticed that we end up opening multiple SMTP connections.

If we exceed max concurrent connections (3 in our case) then it results
in errors and won't be able to send emails.

To avoid this we are refactoring email sending into its own task and
protecting this under a mutex so that only one task be able to send
an email at once.

The task is currently empty but will updated in later commits to
meet the requirements of all tasks that are required to send emails.

Co-authored-by: Le H Ngo <hgle.ngo@gmail.com>