Update manage_inbox() task functions to defer email sending to another task
manage_inbox() task is the one that process mails received from SPIRE
and the reply messages coming from HMRC.

In both cases it updates Mail instances and finally sends out email.
Update relevant functions to defer sending email to another task.

This task receives the serialized Dto object, constructs email
based on the type and sends out email with a locked SMTP connection
so that only one instance of that task is able to send email.
This is to prevent concurrent access to SMTP which can fail if it
exceeds maximum allowed SMTP connections.

Update tests and assertions.