Update rejected licences notify task to defer email sending to a task
If any licences are rejected then we will get a licence reply with
these details. We have a task that notifies users of these rejections
by email.

Refactor existing task to defer email sending to another task so that
we can control concurrent access of SMTP reliably.

Once side benefit is that this task is very simple now, it just builds
a data structure and queues email sending task. It is very unlikely
to fail and hence there is no need of any retry mechanism.

Co-authored-by: Le H Ngo <hgle.ngo@gmail.com>