Update send_email_task() to be generic and just accept the message to send
The main purpose of this task is to send email and it is used by multiple tasks.

Custom serializer is used to serialize arguments send to this task which is
a message.

Update all tasks to adapt to this. The follow up operations are done in
a subtask which is linked to the parent task.