Specify pickle as serializer for send_email_task
The message received to this task is of type MIMEMultipart() which
is not JSON serializable so we are using 'pickle' for this.

Serializer to be used can be specified when scheduling the task as
a kwarg but this can be an issue if the task retries which can
happen in this case. If the SMTP connection is active then the
task won't get lock and will have to retry. If this happens then
it will use json as serializer when the task restarts and fails.

To avoid this specify the serializer for the task using the decorator.
Even after retries it will always use the correct serializer.