Update Mail anonymiser to sanitize exporter details
Previously we were completely overwriting edi_data field with a
static text. This doesn't provide much value if we need to use
this mail for any debug purposes as all details are lost.

PI data mainly present in Exporter and foreign trader details so
it is enough to anonymise those two lines. In this path we
are updating anonymiser to exclude trader details and keep the
rest of them as is.

foreign trader line is handled in subsequent patches.