Add test for mails with invalid edi data
Previously we were replacing this field with a static text which makes
the format invalid. When testing on environments that are loaded
with anonymised data then it is possible to have mails with invalid
edi data so to ensure they are correctly anonymised, add a test to
verify the same.